This PR improves the speed of Upsampling2D on CPU by avoiding for loops.

| | forward | backward |
|-|---------|-----------|
| before | 0.3773 +- 0.0086 sec | 0.1923 +- 0.0210 sec |
| after | 0.0153 +- 0.0004 sec | 0.0110 +- 0.0003 sec |
| speedup | ~24.7 x | ~17.5 x |

The performance measurement script is here: https://gist.github.com/mitmul/b0041b4afd58eaf9489b423cb0345bbb