@niboshi 2017-09-04
Backport?