@mitmul 2017-09-04
I made a backport PR.